Refactor file structure and import paths
The main and test files have been moved to different directories to separate command-line logic from
 the tests. The sorting algorithms are moved to an internal directory to encapsulate them better.
 Import paths have been updated to match the new file structure. This restructuring improves project
  organization and contributes to maintainability.
