+++ internal/randomarray/randomarray.go
